vol: Don't assume enum is an int
The type that an enum is promoted to for arithmetic is compiler
dependent. We can't assume that it's an int, or a short.

Explicitly cast the volume state enum where clang gets upset.

Reviewed-on: http://gerrit.openafs.org/9137
Tested-by: BuildBot <buildbot@rampaginggeek.com>

+++ src/vol/vnode_inline.h
@@ -236,7 +236,7 @@ VnIsErrorState(VnState state)
 static_inline int
 VnIsValidState(VnState state)
 {
-    if ((state >= 0) &&
+    if (((int) state >= 0) &&
 	(state < VN_STATE_COUNT)) {
 	return 1;
     }
